requirements.txt
# Core dependencies
textual>=0.47.0
click>=8.1.0
pyyaml>=6.0

# Search & Data
rapidfuzz>=3.0.0
networkx>=3.0

# Utilities
pydantic>=2.0
rich>=13.0
httpx>=0.24.0

# Performance
orjson>=3.9.0
uvloop>=0.19.0; sys_platform != "win32"
